Add files via upload
Updated graphs to be boxplots (just generally better at representing means) as well as added tests to the ANOVAs (most need minor transformations to be valid but for presentation's sake today we can assume them to all be valid). Still getting to work on tukey tests for ranking so I'll update this again later.