Add files via upload
Updated intro, cleaned up methods - I put an example block for how we can put results in (although I'm yet to figure out how to exclude the resid/pred plots from being printed). You can copy+paste the sections from the final_graphs file I also updated - that top code block is just for importing the data and loading libraries so don't add that to each new plot block. Don't mind the stuff that looks like code in the text area of the results section - just some notes for myself when I go back and address the plots that failed the anova assumptions (though you can ask Faith how necessary that will be for this project/if there's different analyses she recommends because...it's going to be a lot from the looks of it).